network time: take IPv6 addresses into account #7934
While adding this makes sense from a top perspective the use of
get_interface_ip*() let's us know it's using the wild logic to
bind to CARP or IP alias alongside interface defintions.  Normally
the code to do this is pretty small as we only take interfaces,
but there is no point to change this before making an MVC/API
transition here.

Of note is that ntpd wants device names or IP addresses for the
listening directive and then does "something" with it.  How it
reacts to a link-local address is unclear.

Worse still is the fact that "lan_vip" constructs are not properly
handled in the backend writing the configuration, but just going
to leave that as is...